% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_data.R
\docType{data}
\name{dc2010}
\alias{dc2010}
\title{Precinct-level data from the 2010 mayoral election in Washington, DC}
\format{A data frame with 143 rows and 5 variables:
\describe{
  \item{precinct: }{name of the precinct}
  \item{NVoters: }{Number of registed voters}
  \item{NValid: }{Number of valid votes}
  \item{Votes: }{Total votes for Vincent C. Gray}
  \item{a: }{Total abstention}
}}
\usage{
dc2010
}
\description{
This data set contains precinct-level information from the Washington, DC
2010 mayoral election. Only the key information that is required to
estimate fraud in favor of the winner of the elecion (Vincent C. Gray)
is provided in the data set. It includes:
}
\keyword{datasets}
