% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_summary.R
\name{ef_byplot}
\alias{ef_byplot}
\title{Countor plot}
\usage{
ef_byplot(data, samples = NULL, x, y, xlab, ylab,
  contour.fill.color = FALSE, contour.fill.color.key = NULL,
  pch.color.key = NULL, pch.size = 2, pch.show = TRUE,
  legend.title = "")
}
\arguments{
\item{data}{a data frame}

\item{samples}{the output of the function \code{eforensics}. It is optional. If provided, the plot will display the classification of the points in each fraud distribution}

\item{x}{a string with the name of the variable in \code{data} that will be plotted in the x-axis}

\item{y}{a string with the name of the variable in \code{data} that will be plotted in the y-axis}

\item{xlab}{a string with the text to appear in the x-axis}

\item{ylab}{a string with the text to appear in the y-axis}

\item{contour.fill.color}{boolean, if \code{TRUE}, it uses the contour.fill.color.key to color the contour of the bivariate density}

\item{contour.fill.color.key}{a string with the name of the variable in \code{data} to use as a color key of the contour of the bivariate density. It is used only if \code{samples} is not provided}

\item{pch.color.key}{a string with the name of the variable in \code{data} that will be used to color the points. Default is \code{NULL}. It is used only if \code{samples} is not provided}

\item{pch.size}{a integer with the size of the points. Default is 1}

\item{pch.show}{a boolean value indicating if the points should be displayed in the plot (T) or not (F). Default is TRUE}

\item{legend.title}{a string with the title of the legend}
}
\value{
a ggplot object
}
\description{
Create a contour plot with marginal densities
}
