% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_summary.R
\name{ef_classify}
\alias{ef_classify}
\title{Classify observations into fraud distributions}
\usage{
ef_classify(data, samples)
}
\arguments{
\item{data}{a data frame with the data}

\item{samples}{the output of the function \code{eforensics}}
}
\description{
This function classifies the data points into one of three distributions: incremental fraud, extreme fraud, and no fraud.
}
