% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_summary.R
\name{ef_get_local_fraud}
\alias{ef_get_local_fraud}
\title{Compute Fraud Distribution}
\usage{
ef_get_local_fraud(data, samples, model, elegible.voters = NULL,
  sim_data = NULL)
}
\arguments{
\item{data}{the data set used in the estimation}

\item{samples}{the output of \link{eforensics} function}

\item{model}{a string with the model used in the estimation (e.g., 'rn', 'bl'). See \link{eforensics} function}

\item{elegible.voters}{either \code{NULL} (default) or a string with the name of the column containing the number of elegible voters}

\item{sim_data}{the output of the function \link{ef_simulateData}. It is used only if the fraud was estimated using simulated data. If provided, the plot also display the true fraud value. Default \code{NULL}}
}
\value{
It returns a tibble data.frame with the data points classified into no fraud, incremental, or extreme fraud distribution along with the posterior distribution of fraud for each observation
}
\description{
This function returns the estimated posterior distribution of frauds at the individual-level observation
}
