% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_summary.R
\name{ef_get_parameters}
\alias{ef_get_parameters}
\title{Get parameters}
\usage{
ef_get_parameters(samples)
}
\arguments{
\item{samples}{the output of the function \code{eforensics}}
}
\description{
This function returns the parameters of the model
}
