% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_summary.R
\name{ef_get_true}
\alias{ef_get_true}
\title{Get true values from simulated data}
\usage{
ef_get_true(sim_data)
}
\arguments{
\item{sim_data}{an object from \code{ef_simulateData}}
}
\description{
Get true values from simulated data
}
