% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_summary.R
\name{ef_plot}
\alias{ef_plot}
\title{Plot summary of posterior distribution}
\usage{
ef_plot(samples, true = NULL, parse = FALSE,
  plots = c("Abstention and Vote", "Local Fraud Probabilities",
  "Fraud Probability"), title = NULL, subtitle = NULL, xlab = NULL,
  ylab = NULL)
}
\arguments{
\item{true}{used when data was generated by simulation produced by \code{\link{ef_simulateData}} function. The value must the be output of the function \code{\link{ef_get_true()}}}

\item{parse}{boolean, if \code{TRUE}, text in the tick marks of the y-axis is parsed.}

\item{plots}{a string vector with any combination of "Abstention and Vote", "Local Fraud Probabilities", and "Fraud Probability". It defines with plot will be included in the figure. Default includes all.}

\item{title}{a string with the title of the plot}

\item{subtitle}{a string wiht the subtitle of the plot}

\item{xlab}{a string with the lable of the x-axis}

\item{ylab}{a string with the lable of the y-axis}
}
\description{
Plot summary of the samples from posterior distribution produced by \code{\link{eforensics}} function
}
