% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_summary.R
\name{summary.eforensics}
\alias{summary.eforensics}
\title{Summary}
\usage{
\method{summary}{eforensics}(object, ...)
}
\arguments{
\item{object}{the output of the function \code{eforensics}}

\item{...}{join.chains=TRUE can be used to provide summaries of chains after they are combined together}
}
\description{
This function return summaries of the posterior distribution estimated by the function \code{eforensics}
}
