% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_simulate_data.R
\name{ef_simulateData}
\alias{ef_simulateData}
\title{Simulate Election data}
\usage{
ef_simulateData(n = 2000, nCov = 0, nCov.fraud = 0, model,
  pi = NULL, overdispersion = 10)
}
\arguments{
\item{n}{integer, the number of sample points representing observations of election units (e.g., precinct, ballot boxes)}

\item{nCov}{number of covariates affecting both turnout and votes for the winner in each election unit}

\item{nCov.fraud}{number of covariates affecting the occurrence of frauds (either incremental or extreme) in both turnout and votes for the winner}

\item{model}{string which dictates the model from which to simulate data.  Select from one of the following models:
\describe{
  \item{bl}{Simulate data from the binomial-logistic model of election frauds.  Data returned are the }
  \item{bl_none}{Simulate data from the binomial-logistic model of election frauds with no covariates on the frauds.}
  \item{rn_no_alpha}{Simulate data from the restricted normal model with no alpha.}
  \item{rn}{Simulate data from the restricted normal model.}
  \item{rn_wcounts}{Simulate data from the restricted normal model of election frauds and return counts instead of proportions.}
  \item{bbl}{Simulate data from the beta-binomial-logistc model of election frauds.  Simulates data from randomly overdispersed counts.  Overdispersion is a value greater than zero that dictates the amount of overdispersion.  Smaller values for \code{overdispersion} lead to larger amounts of random dispersion within the simulated data.}
}}

\item{pi}{vector with three numbers between 0 and 1 whose sum must add up to 1. If \code{NULL} (default) this vector will be randomly generated}

\item{overdispersion}{numeric, degree of overdispersion of the distributions.  Higher values are less overdispersed.}
}
\value{
The function returns a list with a data frame with the simulated data and a sublist with the parameters used to generate the data:
\describe{
  \item{parameters}{A named vector with the true values used to simulate the data.  Values include the number of observations, values for the mixing parameters, and true values for the intercepts and coefficients associated with the covariates for each of the six no fraud and fraud compoenents of the model.}
  \item{latent}{A named list with the true values of latent quantities used to generate the data.  These are simulated at the observation level.  These values include:}
    \itemize{
      \item{\code{z} The true classification for each observation.  1 is a non-fraudulent observation. 2 is an incrementally fraudulent observation.  And 3 is an extremely fraudulent observation.}
      \item{\code{tau} The proportion of turnout for each observation.  This is associated with non-fraudulent outcomes.}
      \item{\code{nu}  The proportion of votes for the winner for each observation.  This is associated with non-fraudulent outcomes.}
      \item{\code{iota.m} Conditional on being classified as incrementally fraudulent, the proportion of fraudulent turnout.}
      \item{\code{iota.m} Conditional on being classified as incrementally fraudulent, the proportion of fraudulent votes for the winner.}
      \item{\code{chi.m} Conditional on being classified as extremely fraudulent, the proportion of fraudulent turnout.}
      \item{\code{chi.s} Conditional on being classified as extremely fraudulent, the proportion of fraudulent votes for the winner.} 
    }
  \item{data}{A data.frame that includes the simulated data under the defined arguments.  This data.frame may include:}
    \itemize{
      \item{\code{w} The number or proportion of votes for the winner at each observation.}
      \item{\code{a} The number or proportion of abstentions at each observation.}
      \item{\code{x} Covariates.  Each covariate is indexed by the parameter it is associated with in the model and the number.  For example, \code{x1.w} is the first covariate associated with the proportion of non-fraudulent votes for the winner.}
      \item{\code{N} If data is generated by a count model, this is the number of eligible voters at each observation.}
    }
}
}
\description{
This function simulates data sets from the Election Forensics Finite Mixture models.
}
